requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
